+++ src/openloco/input.cpp
@@ -6,6 +6,7 @@ using namespace openloco::interop;
 
 namespace openloco::input
 {
+    loco_global<uint8_t, 0x00508F18> _keyModifier;
     loco_global<uint32_t, 0x00523368> _flags;
     loco_global<uint8_t, 0x0052336D> _state;
     static int32_t _cursor_drag_start_x;
@@ -30,6 +31,11 @@ namespace openloco::input
     input_state state()
     {
         return (input_state)*_state;
+    }
+
+    bool has_key_modifier(key_modifier modifier)
+    {
+        return _keyModifier;
     }
 
     // 0x004BE92A